requirements.txt
fastapi==0.135.3
pandas==3.0.2 
uvicorn==0.42.0
python-multipart==0.0.22
SQLAlchemy==2.0.48
matplotlib==3.10.8
email-validator==2.0.0
google-auth==2.35.0
requests==2.32.3
python-dotenv==1.0.1